+++ src/components/header/controls.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as CSSModules from 'react-css-modules';
-import {BookmarkPane} from './bookmark';
+import {BookmarkPane} from '../bookmark';
 import * as styles from './controls.scss';
 import { UndoRedo } from './undo-redo';
